requirements.txt
pillow
opencv-python
numpy
requests
python-pptx

# OCR（可选，但不装就无法使用OCR功能）
# CPU（推荐稳定）：paddlepaddle 3.x + paddleocr 3.x
# Windows GPU 目前不支持
#   pip install "paddleocr<3"
paddleocr

# PDF导入（可选）
PyMuPDF

# SVG import (optional): direct .svg -> layer
# cairosvg

# 注意：paddlepaddle 在不同平台/CPU/GPU 安装方式不同，建议按官方指引安装：
# https://www.paddlepaddle.org.cn/install/quick
